fastapi_jwt_login/requirements.txt
bcrypt==4.0.1
fastapi
uvicorn
pyjwt==2.10.1
passlib==1.7.4
python-jose==3.5.0
fastapi-csrf-protect==1.0.6
SQLAlchemy==2.0.43